feat: Add support for atoms & react integration
Atoms are a feature modelled on the likes of Jotai and Recoil.

In contrast to Recoil and Jotai, this atom state library supports usage across React roots and similar. It is possible to use this as a reactive global state solution for any library - and you do not have to use React, it just happens to be the library / framework I added first class support for.

Preact is also 100% supported - and just so happens to be the library these atoms were developed for.