feat: Add support for skipWhile predicate in pipe
The skipWhile pipe function works the same way as in RxJS. While the
predicate is true, the Observable result does not emit values.

In order to support this, the process functions now receives the
previously emitted value as an optional second argument. The previous
value is undefined until the first event is emitted through the subject.

BehaviorSubjects will correctly emit its default value immediately to
pipe functions, but bear in mind that the type is narrowed to
IObservable<T> after the first pipe, and this behavior is not present on
IObservable<T>.